pyproject.toml
[tool.poetry]
name = "time-planer"
version = "0.1.0"
description = ""
authors = ["Andy <andy@freilandkiwis.de>"]
readme = "README.md"
packages = [{include = "time_planer"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
pandas = "^1.5.2"
django-colorfield = "^0.8.0"
icecream = "^2.1.3"
pypugjs = "^5.9.12"
django-compressor = "^4.3"
pendulum = "^2.1.2"
django-jazzmin = "^2.6.0"
ipdb = "^0.13.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
